pyproject.toml
[project]
name = "oxmaint-predictive-agent"
version = "0.1.0"
description = "Minimal Predictive Agent for Centrifugal Pump Health — Context Fusion + RAG + XAI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # --- API & Serving ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # --- Agent / Orchestration ---
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    # --- Data Processing ---
    "polars>=1.0.0",
    # --- ML / Numerical Core ---
    "lightgbm>=4.5.0",
    "shap>=0.46.0",
    "scikit-learn>=1.5.0",
    "numpy>=1.26.0",
    # --- Embeddings (BGE-M3) ---
    "sentence-transformers>=3.0.0",
    # --- Vision / SLM (Phi-4 Multimodal) ---
    "transformers>=4.44.0",
    "torch>=2.3.0",
    "Pillow>=10.0.0",
    # --- Database (Supabase + pgvector) ---
    "supabase>=2.0.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "pgvector>=0.3.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    # --- Frontend ---
    "streamlit>=1.38.0",
    "plotly>=5.22.0",
    # --- Validation & Settings ---
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    # --- Utilities ---
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "llama-parse>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py311"
line-length = 100
